refactor(repo): move urlboxoptions tests to own file
